Renamed files to improve cmd semantics
