+++ cmd/tools/root.go
@@ -6,6 +6,7 @@ import (
 	"github.com/dolittle/platform-api/cmd/tools/explore"
 	"github.com/dolittle/platform-api/cmd/tools/job"
 	"github.com/dolittle/platform-api/cmd/tools/m3connector"
+	"github.com/dolittle/platform-api/cmd/tools/microservice"
 	"github.com/dolittle/platform-api/cmd/tools/studio"
 	"github.com/dolittle/platform-api/cmd/tools/terraform"
 	"github.com/dolittle/platform-api/cmd/tools/users"
@@ -29,4 +30,5 @@ func init() {
 	RootCmd.AddCommand(users.RootCMD)
 	RootCmd.AddCommand(application.RootCMD)
 	RootCmd.AddCommand(m3connector.RootCMD)
+	RootCmd.AddCommand(microservice.RootCMD)
 }
